+++ TFG2022Server/Models/LineaPedidoModel.cs
@@ -6,6 +6,6 @@
         public int PedidoLineaPedido { get; set; }
         public int ProductoLineaPedido { get; set; }
         public int Cantidad { get; set; }
-        public int PrecioFinal { get; set; }
+        public double PrecioFinal { get; set; }
     }
 }
